test refactor:
created util functions for generating test lists and items returning
maps of needed ids in the desired formats.

used new functions in tests and referenced the keys in the returned
maps.  this is a lot more readable and changing the logic can be done
in a single place now as long as the data returned is a map which
contains at least the expected keys.