Logging (#24)
* first pass at logging:

testing sending logs to sentry; plan is to log locally in dev and log
to sentry in prod.

still need to config the log levels etc.

* all create update and delete operations logged including errors

* added handling of conditional logging:

this is controled by profiles.clj and ENV VARS
this allows setting the default reported logging level and the
appender (i.e. println vs sentry)

* logging update:

updated the docs to reflect new env vars related to logging.

set sane defaults in the config namespace for logging to reduce manual
steps for default values.